+++ front/src/common/components/ActionButton.tsx
@@ -6,7 +6,7 @@ import React, {
   useContext,
 } from "react";
 import styles from "./ActionButton.module.scss";
-
+import { Link } from "react-router-dom";
 export const ActionButtonContext = createContext<{
   size: "normal" | "small";
 }>({
@@ -35,5 +35,26 @@ export default function ActionButton({
       <span className={styles.ActionButtonIcon}>{icon}</span>
       <span className={styles.ActionButtonContent}>{children}</span>
     </button>
+  );
+}
+
+interface ActionLinkProps extends HTMLAttributes<HTMLLinkElement> {
+  icon: ReactNode;
+  children: ReactNode;
+  to: any;
+}
+export function ActionLink({ icon, children, to }: ActionLinkProps) {
+  const { size } = useContext(ActionButtonContext);
+
+  return (
+    <Link
+      to={to}
+      className={classNames("btn btn--primary", styles.ActionButton, {
+        [styles.ActionButtonSmall]: size === "small",
+      })}
+    >
+      <span className={styles.ActionButtonIcon}>{icon}</span>
+      <span className={styles.ActionButtonContent}>{children}</span>
+    </Link>
   );
 }