+++ front/src/dashboard/components/BSDList/BSDasri/WorkflowAction/utils.ts
@@ -1,19 +1,36 @@
 import * as yup from "yup";
+import { BsdasriQuantity, Bsdasri } from "generated/graphql/types";
 
-import { Bsdasri } from "generated/graphql/types";
-
-import { gql } from "@apollo/client";
-
-export const validationSchema = (form: Bsdasri) =>
+export const signatureValidationSchema = (form: Bsdasri) =>
   yup.object({
-    author: yup.string().nullable().required("Le nom du signataire est requis"),
+    signature: yup.object({
+      author: yup
+        .string()
+        .nullable()
+        .required("Le nom du signataire est requis"),
+    }),
   });
 
-export const SIGN_BSDASRI = gql`
-  mutation SignBsdasri($id: ID!, $input: BsdasriSignatureInput!) {
-    signBsdasri(id: $id, input: $input) {
-      id
-      status
-    }
+export const emissionSignatureSecretCodeValidationSchema = yup.object({
+  signature: yup.object({
+    author: yup.string().nullable().required("Le nom du signataire est requis"),
+    securityCode: yup.number().required("Le code de signature est obligatoire"),
+  }),
+});
+
+export const getInitialQuantityFn = (quantity?: BsdasriQuantity | null) => ({
+  value: quantity?.value,
+  type: quantity?.type,
+});
+
+export const prefillWasteDetails = dasri => {
+  if (!dasri?.transport?.wasteDetails?.packagingInfos?.length) {
+    dasri.transport.wasteDetails.packagingInfos =
+      dasri?.emission?.wasteDetails?.packagingInfos;
   }
-`;
+  if (!dasri?.reception?.wasteDetails?.packagingInfos?.length) {
+    dasri.reception.wasteDetails.packagingInfos =
+      dasri?.transport?.wasteDetails?.packagingInfos;
+  }
+  return dasri;
+};